requirements.txt
Flask
python-dotenv
PyPDF2
python-docx
requests
duckduckgo-search
pandas
python-pptx
pytesseract
Pillow
SpeechRecognition
openpyxl